requirements.txt
Babel==2.9.1
Bottleneck==1.4.0
Brotli==1.1.0
Cython==3.0.10
GitPython==3.1.43
Jinja2==3.1.4
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
PySocks==1.7.1
PyYAML==6.0.1
Pygments==2.18.0
SQLAlchemy==2.0.31
Werkzeug==3.0.3
absl-py==2.1.0
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.13.2
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
black==24.4.2
blinker==1.8.2
cachetools==5.4.0
catboost==1.2.5
category-encoders==2.6.3
certifi==2024.7.4
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmdstanpy==1.2.4
colorlog==6.8.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
efficientnet-pytorch==0.7.1
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.15.4
filetype==1.2.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
fvcore==0.1.5.post20221221
gitdb==4.0.11
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.65.2
holidays==0.53
huggingface-hub==0.24.5
hydra-core==1.3.2
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
iopath==0.1.9
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
jedi==0.19.1
joblib==1.4.2
jsonschema-specifications==2023.12.1
jsonschema==4.23.0
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lightgbm==4.5.0
lightning-utilities==0.11.6
magic-filter==1.0.12
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
matplotlib==3.9.1
mdurl==0.1.2
mplcyberpunk==0.7.1
mpmath==1.3.0
multidict==6.0.5
munch==4.0.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numexpr==2.10.1
numpy
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python-headless
optuna==3.6.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
pmdarima==2.0.4
portalocker==2.10.1
pretrainedmodels==0.7.4
prompt_toolkit==3.0.47
prophet==1.1.5
protobuf==4.25.4
protobuf
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pycocotools==2.0.8
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.3
referencing==0.35.1
requests-toolbelt==1.0.0
requests==2.32.3
rich==13.7.1
roboflow==1.1.37
rpds-py==0.19.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
segmentation-models-pytorch==0.3.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
stanio==0.5.1
statsmodels==0.14.2
streamlit==1.37.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard-data-server==0.7.2
tensorboard==2.17.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
timm==0.9.2
toml==0.10.2
toolz==0.12.1
torch==2.4.0
torchmetrics==1.4.0.post0
torchutils==0.0.4
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics-thop==2.0.0
ultralytics==8.2.70
unicodedata2==15.1.0
urllib3==2.2.2
watchdog==4.0.1
wcwidth==0.2.13
widgetsnbextension==4.0.11
xgboost==2.1.0
xlrd==2.0.1
yacs==0.1.8
yarl==1.9.4
zipp==3.19.2
